# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/profile": {
      "get": {
        "description": "Returns the User record with the nested UserProfile.",
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "uuid-v4"
                        },
                        "firebaseUid": {
                          "type": "string",
                          "example": "firebase-uid"
                        },
                        "email": {
                          "type": "string",
                          "example": "user@example.com",
                          "nullable": true
                        },
                        "phoneNumber": {
                          "type": "string",
                          "nullable": true
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "profile": {
                          "$ref": "#/components/schemas/UserProfileDto"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      },
      "put": {
        "description": "Updates name, gender, and date of birth. All fields optional — send only what changed.",
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserProfileDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update personal info",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/profile/physical": {
      "put": {
        "description": "Updates height (cm) and weight (kg). Used during onboarding and settings.",
        "operationId": "UsersController_updatePhysical",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePhysicalDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Physical info updated",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserProfileDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update physical measurements",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/users/onboarding-status": {
      "get": {
        "description": "`hasProfile` = name/gender filled. `hasPhysicalData` = height/weight filled. `onboardingCompleted` is set to true only after user explicitly completes the flow.",
        "operationId": "UsersController_getOnboardingStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Onboarding status",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "onboardingCompleted": {
                          "type": "boolean",
                          "example": false
                        },
                        "hasProfile": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPhysicalData": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get onboarding completion status",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/progress/daily": {
      "get": {
        "description": "Returns how much the user actually exercised, was upright, and used the device today. Combine with GET /api/goals/daily to show progress bars.",
        "operationId": "ProgressController_getDailyProgress",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daily progress retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/DailyProgressResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get today's actual progress",
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/progress/analytics": {
      "get": {
        "description": "Returns aggregated stats and daily breakdown for the chosen period. Use `dailyData` to draw charts.",
        "operationId": "ProgressController_getAnalytics",
        "parameters": [
          {
            "name": "period",
            "required": false,
            "in": "query",
            "description": "Time period to aggregate. Defaults to \"week\".",
            "schema": {
              "enum": [
                "day",
                "week",
                "month",
                "year"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Analytics retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/AnalyticsResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get analytics summary for a time period",
        "tags": [
          "Progress"
        ]
      }
    },
    "/api/progress/streak": {
      "get": {
        "description": "Returns how many consecutive days the user has been active.",
        "operationId": "ProgressController_getStreak",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Streak retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/StreakResponseDto"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current activity streak",
        "tags": [
          "Progress"
        ]
      }
    },
...
  },
  "info": {
    "title": "Straight API",
    "description": "Backend API for the Straight posture correction mobile app. All endpoints except /api/users/sync require a Firebase Bearer token.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "userId": {
            "type": "string",
            "example": "uuid-v4"
          },
          "firstName": {
            "type": "object",
            "example": "Ayşe"
          },
          "lastName": {
            "type": "object",
            "example": "Yılmaz"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "NOT_SPECIFIED"
            ],
            "example": "FEMALE"
          },
          "dateOfBirth": {
            "type": "object",
            "example": "1995-06-15T00:00:00.000Z"
          },
          "heightCm": {
            "type": "object",
            "example": 168
          },
          "weightKg": {
            "type": "object",
            "example": 65.5
          },
          "onboardingCompleted": {
            "type": "boolean",
            "example": false
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "userId",
          "gender",
          "onboardingCompleted",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateProfileDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "example": "Ayşe"
          },
          "lastName": {
            "type": "string",
            "example": "Yılmaz"
          },
          "gender": {
            "type": "string",
            "enum": [
              "MALE",
              "FEMALE",
              "NOT_SPECIFIED"
            ],
            "example": "FEMALE"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "1995-06-15"
          }
        }
      },
      "UpdatePhysicalDto": {
        "type": "object",
        "properties": {
          "heightCm": {
            "type": "number",
            "example": 168,
            "description": "Height in centimeters"
          },
          "weightKg": {
            "type": "number",
            "example": 65.5,
            "description": "Weight in kilograms"
          }
        }
      },
...
      "DailyProgressResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "uuid-v4"
          },
          "userId": {
            "type": "string",
            "example": "uuid-v4"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T00:00:00.000Z"
          },
          "exerciseActualMinutes": {
            "type": "number",
            "example": 28.5,
            "description": "Minutes exercised today"
          },
          "uprightActualMinutes": {
            "type": "number",
            "example": 320,
            "description": "Minutes spent upright today"
          },
          "deviceUsageActualMinutes": {
            "type": "number",
            "example": 55,
            "description": "Minutes device was used today"
          },
          "streakDays": {
            "type": "number",
            "example": 7,
            "description": "Consecutive days streak"
          }
        },
        "required": [
          "date",
          "exerciseActualMinutes",
          "uprightActualMinutes",
          "deviceUsageActualMinutes",
          "streakDays"
        ]
      },
      "DailyProgressDataDto": {
        "type": "object",
        "properties": {
          "date": {
            "format": "date-time",
            "type": "string"
          },
          "exerciseActualMinutes": {
            "type": "number",
            "example": 28.5
          },
          "uprightActualMinutes": {
            "type": "number",
            "example": 320
          },
          "deviceUsageActualMinutes": {
            "type": "number",
            "example": 55
          },
          "streakDays": {
            "type": "number",
            "example": 7
          }
        },
        "required": [
          "date",
          "exerciseActualMinutes",
          "uprightActualMinutes",
          "deviceUsageActualMinutes",
          "streakDays"
        ]
      },
      "AnalyticsResponseDto": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string",
            "enum": [
              "day",
              "week",
              "month",
              "year"
            ],
            "example": "week"
          },
          "startDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-08T00:00:00.000Z"
          },
          "endDate": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T00:00:00.000Z"
          },
          "totalSessions": {
            "type": "number",
            "example": 12,
            "description": "Total sessions in period"
          },
          "averageUprightMinutes": {
            "type": "number",
            "example": 245,
            "description": "Average upright minutes per day"
          },
          "averageScore": {
            "type": "object",
            "example": 79,
            "description": "Average posture score, null if no sessions"
          },
          "dailyData": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DailyProgressDataDto"
            }
          }
        },
        "required": [
          "period",
          "startDate",
          "endDate",
          "totalSessions",
          "averageUprightMinutes",
          "dailyData"
        ]
      },
      "StreakResponseDto": {
        "type": "object",
        "properties": {
          "currentStreak": {
            "type": "number",
            "example": 7,
            "description": "Number of consecutive active days"
          },
          "date": {
            "format": "date-time",
            "type": "string",
            "example": "2024-01-15T00:00:00.000Z"
          }
        },
        "required": [
          "currentStreak",
          "date"
        ]
      },
...
    }
  }
}